[CAIRO] Make GLContextGLX a subclass of GLContext https://bugs.webkit.org/show_bug.cgi?id=81285
Reviewed by Gustavo Noronha Silva.

.:

* GNUmakefile.am: Add a configuration option for activating GLX.
This is selected automatically now, but in the future there will be
the choice to turn on EGL and turn off GLX at compilation time.
* configure.ac: Ditto.

Source/WebCore:

No new tests. This should not change behavior.

* GNUmakefile.list.am: Add new GLContextGLX files.
* platform/graphics/cairo/GLContext.cpp: Added. Composed of the
EGL/GLX independent portions of GLContext. Made the current context
a static variable so we can keep getting the current context platform-
-independent.
* platform/graphics/cairo/GLContext.h: Make this class an abstract
class. GLContextGLX is a concrete implementation.
* platform/graphics/cairo/GraphicsContext3DPrivate.cpp: Updated to reflect
slightly different calling conventions to create a sharing context.
* platform/graphics/glx/GLContextGLX.cpp: Updated to reflect
that this code is in a subclass now.
* platform/graphics/glx/GLContextGLX.h: Copied from Source/WebCore/platform/graphics/cairo/GLContext.h.
* platform/graphics/gtk/GLContextGtk.cpp:
(WebCore::GLContext::getContextForWidget): Updated to respect the new
USE(GLX) flag.

Identifier: 101204@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@113909 268f45cc-cd09-0410-ab3c-d52691b4dbfc