[Qt][WK2] Minibrowser asserts on startup at QtViewportInteractionEngine::ensureContentWithinViewportBoundary() https://bugs.webkit.org/show_bug.cgi?id=84172
Patch by Zalan Bujtas <zbujtas@gmail.com> on 2012-04-18
Reviewed by Simon Hausmann.

After the viewport computing refactor, ensureContentWithinViewportBoundary() can be called
with resumed page. QtViewportInteractionEngine::ensureContentWithinViewportBoundary()
functionality does not necessarily require the page to be suspended. It's the caller's context
that determines whether the page needs to be suspended, so it's the caller's responsibility
to enforce the ASSERT().
Remove the invalid ASSERT on m_suspendCount at QtViewportInteractionEngine::ensureContentWithinViewportBoundary()

* UIProcess/qt/QtViewportInteractionEngine.cpp:
(WebKit::QtViewportInteractionEngine::ensureContentWithinViewportBoundary):

Identifier: 101744@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114496 268f45cc-cd09-0410-ab3c-d52691b4dbfc