Inserting empty html moves caret. https://bugs.webkit.org/show_bug.cgi?id=71771
Patch by Antaryami Pandia <antaryami.pandia@motorola.com> on 2012-04-19
Reviewed by Ryosuke Niwa.

Source/WebCore:

When we place the cursor in the middle of a text node and try to insert some text
between, then we split text node. But in this case we have nothing to insert since
the string to be inserted is empty. So the check for fragments should precedes the
call to code block containing splitTextNode.

Test: editing/inserting/insert-empty-html.html

* editing/ReplaceSelectionCommand.cpp:
(WebCore::ReplaceSelectionCommand::doApply):

LayoutTests:

Test inserting empty html.

* editing/inserting/insert-empty-html-expected.txt: Added.
* editing/inserting/insert-empty-html.html: Added.

Identifier: 101914@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114718 268f45cc-cd09-0410-ab3c-d52691b4dbfc