Layout Test svg/transforms/transform-origin-css-property.xhtml is failing https://bugs.webkit.org/show_bug.cgi?id=84203
Reviewed by Simon Fraser.

This test compares CSS div's n a ref-test with a single spec for transform-origin
to SVG rects with multipe specs for transform-origin that are supposed
to all be equivalent, such as [50%,50%] and [center, center]. It fails
due to one-pixel differences on Chromium.

This change renders as many divs in the ref-test as SVG rects in the
test, and ends up with the same result. It seems likely that one of
the several specs that are used is computed slightly differently
(maybe 50% is not the same as center).

* platform/chromium/test_expectations.txt:
* svg/transforms/transform-origin-css-property-expected.xhtml:


Identifier: 101948@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114752 268f45cc-cd09-0410-ab3c-d52691b4dbfc