[chromium] Remove special case for recreating layers during sync https://bugs.webkit.org/show_bug.cgi?id=84458
Reviewed by James Robinson.

Source/WebCore:

We no longer delete layers on the impl side, unless it is due to
deleting the layer on webkit side. So we will no longer have a
situation where we need to push properties from an old
LayerChromium to a new CCLayerImpl.

* platform/graphics/chromium/TreeSynchronizer.cpp:
(WebCore::TreeSynchronizer::reuseOrCreateCCLayerImpl):

Source/WebKit/chromium:

* tests/TreeSynchronizerTest.cpp:


Identifier: 101999@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114808 268f45cc-cd09-0410-ab3c-d52691b4dbfc