REGRESSION(r114870): Assertion failure in ElementAttributeData::setAttr(). <http://webkit.org/b/84581>
Reviewed by Antti Koivisto.

Attach the Attr and bump m_attrCount manually in ensureAttr() instead of calling
setAttr(). The latter asserts that the Attr isn't present in the map, which
obviously isn't true after we've just added it.

This has the added effect of removing one unnecessary hash lookup.

* dom/ElementAttributeData.cpp:
(WebCore::ElementAttributeData::ensureAttr):

Identifier: 102068@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@114883 268f45cc-cd09-0410-ab3c-d52691b4dbfc