Biolab disaster crashes on ToT https://bugs.webkit.org/show_bug.cgi?id=84898
Reviewed by Filip Pizlo.

I recently added an assertion to the Interpreter to catch incorrect
updates of topCallFrame.  This caused a bunch of sites (including biolab
disaster) to crash as we were not correctly handling callee registers
of inlined functions, leading to a mismatch.

I could not actually make this trigger directly, although it does trigger
already on some of the GTK and QT bots.

* dfg/DFGVirtualRegisterAllocationPhase.cpp:
(JSC::DFG::VirtualRegisterAllocationPhase::run):

Identifier: 102404@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115252 268f45cc-cd09-0410-ab3c-d52691b4dbfc