Use WebKit types for the cache of ObjcClass::methodsNamed() https://bugs.webkit.org/show_bug.cgi?id=85012
Patch by Benjamin Poulain <bpoulain@apple.com> on 2012-04-26
Reviewed by Geoffrey Garen.

This patch redefines the method cache ObjcClass to avoid memory allocations in the case of positive match.

Instead of using the converted name as the key, the original identifier string is used. This shortcuts
all the other operations when there is a match.

A side effect is a method can appear multiple times in the cache if it is invoked with different names using
the escape character "$". An attaquer could bloat the cache with a few hundreds strings.
In the common case, having each name mapped is an improvment.

* bridge/objc/objc_class.h:
(ObjcClass):
* bridge/objc/objc_class.mm:
(JSC::Bindings::ObjcClass::ObjcClass):
(JSC::Bindings::ObjcClass::methodsNamed):


Identifier: 102530@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115392 268f45cc-cd09-0410-ab3c-d52691b4dbfc