Generalize the single character optimization of numberProtoFuncToString https://bugs.webkit.org/show_bug.cgi?id=85027
Reviewed by Geoffrey Garen.

The function numberProtoFuncToString() has an optimization to use SmallStrings::singleCharacterString()
when the radix is 36.

This patch generalize the optimization for any radix. Any positive number smaller than its radix
can be represented by a single character of radixDigits.

This makes numberProtoFuncToString() about twice as fast for this case of single digit conversion.

* runtime/NumberPrototype.cpp:
(JSC::numberProtoFuncToString):


Identifier: 102598@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115465 268f45cc-cd09-0410-ab3c-d52691b4dbfc