Add a barrier-style dispatch member function to ScrollingThread https://bugs.webkit.org/show_bug.cgi?id=85228
Reviewed by Sam Weinig.

Add a ScrollingThread::dispatchBarrier function which takes a WTF::Function and dispatches it to the main thread
once all the currently scheduled scrolling thread functions have run. This is to be used for synchronization between the
scrolling thread and the main thread.

* page/scrolling/ScrollingThread.cpp:
(WebCore::callFunctionOnMainThread):
(WebCore):
(WebCore::ScrollingThread::dispatchBarrier):
* page/scrolling/ScrollingThread.h:
(ScrollingThread):

Identifier: 102768@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115665 268f45cc-cd09-0410-ab3c-d52691b4dbfc