Fullscreen pop-up logic restored to using processingUserGesture. https://bugs.webkit.org/show_bug.cgi?id=85105
WebKit was recently updated to the new Fullscreen API:
http://dvcs.w3.org/hg/fullscreen/raw-file/tip/Overview.html#api
http://trac.webkit.org/changeset/111028

This change reverts back to using processingUserGesture() instead
of DOMWindow::allowPopUp(). This fixes incorrect behavior in
at least the Chromium port and is consistent with the cited
definition of "allowed to show a pop-up":
  An algorithm is allowed to show a pop-up if, in the task in which the algorithm is running, either:
  - an activation behavior is currently being processed whose click event was trusted, or
  - the event listener for a trusted click event is being handled.

Reviewed by Dimitri Glazkov.

No new tests.

* dom/Document.cpp:
(WebCore::Document::requestFullScreenForElement):


Identifier: 102876@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115792 268f45cc-cd09-0410-ab3c-d52691b4dbfc