Drag and drop text into table is pasting the text in the next <td> element https://bugs.webkit.org/show_bug.cgi?id=75004
Reviewed by Darin Adler.

The bug was caused by ReplaceSelectionCommand adjusting the insertion position to be before
of the block element containing the insertion position even when the block element is a table cell.

Fixed the bug by not moving the insertion position before the table cell in this case.

Test: editing/pasteboard/paste-into-table-cell.html

* editing/ReplaceSelectionCommand.cpp:
(WebCore::ReplaceSelectionCommand::doApply):


Identifier: 102948@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@115869 268f45cc-cd09-0410-ab3c-d52691b4dbfc